requirements.txt
SpeechRecognition
pyaudio
google-generativeai
python-dotenv
elevenlabs
pygame
